requirements.txt
streamlit
pyyaml
spaCy
nltk
whisper
gTTS
googletrans==4.0.0-rc1
